Move the shared stake files into stakeForm
The stakeDrawer folder held only pieces stakeForm uses, so the name
was misleading. Move the reducer, submit button and types there and
rename the reducer identifiers from Drawer to Form.
