+++ web/src/pages/earn/components/stakeForm/stakeDepositProgressDrawer.tsx
@@ -15,7 +15,7 @@ import { useTranslation } from "react-i18next";
 import type { TokenWithGateway } from "types";
 import type { Address } from "viem";
 
-import type { DepositStep } from "../stakeDrawer/stakeDrawerReducer";
+import type { DepositStep } from "./stakeFormReducer";
 
 type Props = {
   amount: string;
