+++ web/src/pages/earn/components/stakeForm/types.ts
